Add files via upload
This project aims to develop a system of IoT devices to prevent under-loading / overloading of Railway wagons. Coal India Limited has been supplying coal to its consumers by Rail through Railway Sidings. The railway wagons at such sidings are loaded through contractual means by a payloader. The loading of wagons by contractual arrangements often result in overloading or underloading of the Railway wagons. The rules of      IPC for overloading and underloading are notified by the Railway. Any penalty for overloading charged by the Railway for any consignment is payable by the purchaser. However, in case of underloading of wagons, credit for idle freight is adjusted in coal bills. Thus any idle freight for under-loading is borne by CIL. During 2021-22, the expense for under-loading was nearly Rs.593 Cr. whereas the contract for wagon loading itself was only Rs.276 Cr. Hence, a digital solution in the form of sensor/ IoT is needed to prevent the under-loading and overloading of Railway wagons.